feat(finance): accept string|null in CurrencyFormatter::formatHtml, drop caller casts (#8459)

+++ src/finance/views/dashboard.php
@@ -57,7 +57,7 @@ $sRootPath = SystemURLs::getRootPath();
                             </span>
                         </div>
                         <div class="col">
-                            <div class="fw-medium"><?= CurrencyFormatter::formatHtml((float) ($ytdPaymentTotal ?? 0)) ?></div>
+                            <div class="fw-medium"><?= CurrencyFormatter::formatHtml($ytdPaymentTotal ?? 0) ?></div>
                             <div class="text-body-secondary"><?= gettext('YTD Payments') ?></div>
                         </div>
                     </div>
@@ -74,7 +74,7 @@ $sRootPath = SystemURLs::getRootPath();
                             </span>
                         </div>
                         <div class="col">
-                            <div class="fw-medium"><?= CurrencyFormatter::formatHtml((float) ($ytdPledgeTotal ?? 0)) ?></div>
+                            <div class="fw-medium"><?= CurrencyFormatter::formatHtml($ytdPledgeTotal ?? 0) ?></div>
                             <div class="text-body-secondary"><?= gettext('YTD Pledges') ?></div>
                         </div>
                     </div>
@@ -309,7 +309,7 @@ $sRootPath = SystemURLs::getRootPath();
                                     <td><?= $deposit->getDate('M j, Y') ?></td>
                                     <td><span class="badge bg-blue-lt text-blue"><?= $deposit->getType() ?></span></td>
                                     <td class="text-truncate finance-truncate"><?= InputUtils::escapeHTML($deposit->getComment() ?? '') ?></td>
-                                    <td class="text-end fw-bold"><?= CurrencyFormatter::formatHtml((float) ($deposit->getVirtualColumn('totalAmount') ?? 0)) ?></td>
+                                    <td class="text-end fw-bold"><?= CurrencyFormatter::formatHtml($deposit->getVirtualColumn('totalAmount') ?? 0) ?></td>
                                     <td>
                                         <?php if ($deposit->getClosed()): ?>
                                         <span class="badge bg-green-lt text-green"><?= gettext('Closed') ?></span>
@@ -363,7 +363,7 @@ $sRootPath = SystemURLs::getRootPath();
                     </div>
                     <hr>
                     <div class="text-center mb-3">
-                        <div class="h3 text-success mb-0"><?= CurrencyFormatter::formatHtml((float) ($currentDeposit->getVirtualColumn('totalAmount') ?? 0)) ?></div>
+                        <div class="h3 text-success mb-0"><?= CurrencyFormatter::formatHtml($currentDeposit->getVirtualColumn('totalAmount') ?? 0) ?></div>
                         <small class="text-body-secondary"><?= gettext('Total Amount') ?></small>
                     </div>
                     <a href="<?= $sRootPath ?>/DepositSlipEditor.php?DepositSlipID=<?= $currentDeposit->getId() ?>" class="btn btn-primary w-100">
